+++ test/integration/test_seq_copy_int.py
@@ -142,7 +142,24 @@ ENCODER_DECODER_SETTINGS_TEMPLATE = [
      " --source-factors-num-embed 8 2 8"
      " --target-factors-combine sum --target-factors-share-embedding false"
      " --target-factors-num-embed 8"
-     " --scheduled-sampling full --scheduled-sampling-eval --scheduled-sampling-rate 0.5 --scheduled-sampling-warmup 1",
+     " --scheduled-sampling full --scheduled-sampling-rate 0.5 --scheduled-sampling-rate-warmup 1"
+     " --scheduled-sampling-temperature 2 --scheduled-sampling-temperature-warmup 1",
+     "--beam-size 2",
+     True, 3, 1),
+     # Basic transformer with source and target factors, uniform scheduled sampling
+    ("--encoder transformer --decoder {decoder}"
+     " --num-layers 2 --transformer-attention-heads 2 --transformer-model-size 8 --num-embed 8"
+     " --transformer-feed-forward-num-hidden 16"
+     " --transformer-dropout-prepost 0.1 --transformer-preprocess n --transformer-postprocess dr"
+     " --weight-tying-type trg_softmax"
+     " --batch-size 2 --max-updates 2 --batch-type sentence --decode-and-evaluate 0"
+     " --checkpoint-interval 2 --optimizer adam --initial-learning-rate 0.01"
+     " --source-factors-combine sum concat average --source-factors-share-embedding true false true"
+     " --source-factors-num-embed 8 2 8"
+     " --target-factors-combine sum --target-factors-share-embedding false"
+     " --target-factors-num-embed 8"
+     " --scheduled-sampling uniform --scheduled-sampling-rate 0.5 --scheduled-sampling-rate-warmup 1"
+     " --scheduled-sampling-temperature 2 --scheduled-sampling-temperature-warmup 1",
      "--beam-size 2",
      True, 3, 1),
      # Basic transformer with source and target factors, fast scheduled sampling
@@ -157,10 +174,11 @@ ENCODER_DECODER_SETTINGS_TEMPLATE = [
      " --source-factors-num-embed 8 2 8"
      " --target-factors-combine sum --target-factors-share-embedding false"
      " --target-factors-num-embed 8"
-     " --scheduled-sampling fast --scheduled-sampling-eval --scheduled-sampling-rate 0.5 --scheduled-sampling-warmup 1",
+     " --scheduled-sampling fast --scheduled-sampling-rate 0.5 --scheduled-sampling-rate-warmup 1"
+     " --scheduled-sampling-temperature 2 --scheduled-sampling-temperature-warmup 1",
      "--beam-size 2",
      True, 3, 1),
-      # Basic transformer with source and target factors, uniform scheduled sampling
+      # Basic transformer with source and target factors, diff scheduled sampling
     ("--encoder transformer --decoder {decoder}"
      " --num-layers 2 --transformer-attention-heads 2 --transformer-model-size 8 --num-embed 8"
      " --transformer-feed-forward-num-hidden 16"
@@ -172,8 +190,8 @@ ENCODER_DECODER_SETTINGS_TEMPLATE = [
      " --source-factors-num-embed 8 2 8"
      " --target-factors-combine sum --target-factors-share-embedding false"
      " --target-factors-num-embed 8"
-     " --scheduled-sampling uniform --scheduled-sampling-eval --scheduled-sampling-rate 0.5"
-     " --scheduled-sampling-warmup 1",
+     " --scheduled-sampling diff --scheduled-sampling-rate 0.5 --scheduled-sampling-rate-warmup 1"
+     " --scheduled-sampling-temperature 2 --scheduled-sampling-temperature-warmup 1 --scheduled-sampling-sample",
      "--beam-size 2",
      True, 3, 1),
 ]